requirements.txt
functions-framework==3.*
requests
sendgrid==6.9.2
google-cloud-datastore
google.cloud
